chore: argowf client sturcture updated
