+++ kubewatch/pkg/config/ciConfig.go
@@ -14,20 +14,20 @@
  * limitations under the License.
  */
 
-package bean
+package config
 
-type ClientType string
+import "github.com/caarlos0/env"
 
-const (
-	ArgoCDClientType         ClientType = "ArgoCD"
-	CiArgoWorkflowClientType ClientType = "CiArgoWorkflow"
-	CdArgoWorkflowClientType ClientType = "CdArgoWorkflow"
-	SystemExecutorClientType ClientType = "SystemExecutor"
-)
+type CiConfig struct {
+	// DefaultNamespace is the namespace where all CI workflows are scheduled
+	DefaultNamespace string `env:"DEFAULT_NAMESPACE" envDefault:"devtron-ci"`
 
-var SupportedClientMap = map[ClientType]bool{
-	ArgoCDClientType:         true,
-	CiArgoWorkflowClientType: true,
-	CdArgoWorkflowClientType: true,
-	SystemExecutorClientType: true,
+	// CiInformer is used to determine whether CI informer is enabled or not
+	CiInformer bool `env:"CI_INFORMER" envDefault:"true"`
+}
+
+func getCiConfig() (*CiConfig, error) {
+	ciConfig := &CiConfig{}
+	err := env.Parse(ciConfig)
+	return ciConfig, err
 }
