--- kubewatch/pkg/informer/cluster/argoWf/helper.go
@@ -1,145 +1,0 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package argoWf
-
-import (
-	"fmt"
-	repository "github.com/devtron-labs/kubewatch/pkg/cluster"
-	informerBean "github.com/devtron-labs/kubewatch/pkg/informer/bean"
-	informerErr "github.com/devtron-labs/kubewatch/pkg/informer/errors"
-	"github.com/devtron-labs/kubewatch/pkg/middleware"
-	resourceBean "github.com/devtron-labs/kubewatch/pkg/resource/bean"
-	"golang.org/x/exp/maps"
-	"time"
-)
-
-func (impl *InformerImpl) getCiArgoWfStopper(clusterId int) (*informerBean.SharedStopper, bool) {
-	stopper, ok := impl.argoWfCiInformerStopper[clusterId]
-	if ok {
-		return stopper, stopper.HasInformer()
-	}
-	return stopper, false
-}
-
-func (impl *InformerImpl) getCdArgoWfStopper(clusterId int) (*informerBean.SharedStopper, bool) {
-	stopper, ok := impl.argoWfCiInformerStopper[clusterId]
-	if ok {
-		return stopper, stopper.HasInformer()
-	}
-	return stopper, false
-}
-
-func (impl *InformerImpl) getStoppableCiArgoWfs() []int {
-	return maps.Keys(impl.argoWfCiInformerStopper)
-}
-
-func (impl *InformerImpl) getStoppableCdArgoWfs() []int {
-	return maps.Keys(impl.argoWfCdInformerStopper)
-}
-
-func (impl *InformerImpl) getStoppableClusterIds() []int {
-	stoppableClusterIdsMap := make(map[int]bool)
-	for _, clusterId := range impl.getStoppableCiArgoWfs() {
-		stoppableClusterIdsMap[clusterId] = true
-	}
-	for _, clusterId := range impl.getStoppableCdArgoWfs() {
-		stoppableClusterIdsMap[clusterId] = true
-	}
-	return maps.Keys(stoppableClusterIdsMap)
-}
-
-func (impl *InformerImpl) getCiArgoWfStopChannel(clusterLabels *informerBean.ClusterLabels) (chan struct{}, error) {
-	stopChannel := make(chan struct{})
-	stopper, ok := impl.argoWfCiInformerStopper[clusterLabels.ClusterId]
-	if ok && stopper.HasInformer() {
-		impl.logger.Debug(fmt.Sprintf("ci argo workflow informer for %s already exist", clusterLabels.ClusterName))
-		return stopChannel, informerErr.AlreadyExists
-	}
-	stopper = stopper.GetStopper(stopChannel)
-	impl.argoWfCiInformerStopper[clusterLabels.ClusterId] = stopper
-	return stopChannel, nil
-}
-
-func (impl *InformerImpl) getCdArgoWfStopChannel(clusterLabels *informerBean.ClusterLabels) (chan struct{}, error) {
-	stopChannel := make(chan struct{})
-	stopper, ok := impl.argoWfCdInformerStopper[clusterLabels.ClusterId]
-	if ok && stopper.HasInformer() {
-		impl.logger.Debug(fmt.Sprintf("cd argo workflow informer for %s already exist", clusterLabels.ClusterName))
-		return stopChannel, informerErr.AlreadyExists
-	}
-	stopper = stopper.GetStopper(stopChannel)
-	impl.argoWfCdInformerStopper[clusterLabels.ClusterId] = stopper
-	return stopChannel, nil
-}
-
-func (impl *InformerImpl) startCiArgoWfInformer(clusterInfo *repository.Cluster) error {
-	if !impl.appConfig.GetCiConfig().CiInformer {
-		impl.logger.Debugw("ci argo workflow informer is not enabled, skipping...", "clusterId", clusterInfo.Id, "clusterName", clusterInfo.ClusterName, "appConfig", impl.appConfig)
-		return nil
-	}
-	startTime := time.Now()
-	defer func() {
-		impl.logger.Debugw("time taken to start ci argo workflow informer", "clusterId", clusterInfo.Id, "time", time.Since(startTime))
-	}()
-	restConfig := impl.k8sUtil.GetK8sConfigForCluster(clusterInfo)
-	ciWfInformer := impl.informerClient.GetSharedInformerClient(resourceBean.CiWorkflowResourceType)
-	clusterLabels := informerBean.NewClusterLabels(clusterInfo.ClusterName, clusterInfo.Id)
-	workflowInformer, err := ciWfInformer.GetSharedInformer(clusterInfo.Id, impl.appConfig.GetCiWfNamespace(), restConfig)
-	if err != nil {
-		impl.logger.Errorw("error in starting workflow informer", "err", err)
-		middleware.IncUnregisteredInformers(clusterLabels, middleware.CI_STAGE_ARGO_WORKFLOW)
-	}
-	stopChan, err := impl.getCiArgoWfStopChannel(clusterLabels)
-	if err != nil {
-		return err
-	}
-	runnable := func() {
-		workflowInformer.Run(stopChan)
-		impl.logger.Infow("informer started for ci argo workflow", "clusterId", clusterInfo.Id, "clusterName", clusterInfo.ClusterName)
-	}
-	impl.asyncRunnable.Execute(runnable)
-	return nil
-}
-
-func (impl *InformerImpl) startCdArgoWfInformer(clusterInfo *repository.Cluster) error {
-	if !impl.appConfig.GetCdConfig().CdInformer {
-		impl.logger.Debugw("cd argo workflow informer is not enabled, skipping...", "clusterId", clusterInfo.Id, "clusterName", clusterInfo.ClusterName, "appConfig", impl.appConfig)
-		return nil
-	}
-	startTime := time.Now()
-	defer func() {
-		impl.logger.Debugw("time taken to start cd argo workflow informer", "clusterId", clusterInfo.Id, "time", time.Since(startTime))
-	}()
-	restConfig := impl.k8sUtil.GetK8sConfigForCluster(clusterInfo)
-	cdWfInformer := impl.informerClient.GetSharedInformerClient(resourceBean.CdWorkflowResourceType)
-	clusterLabels := informerBean.NewClusterLabels(clusterInfo.ClusterName, clusterInfo.Id)
-	workflowInformer, err := cdWfInformer.GetSharedInformer(clusterInfo.Id, impl.appConfig.GetCdWfNamespace(), restConfig)
-	if err != nil {
-		impl.logger.Errorw("error in starting workflow informer", "err", err)
-		middleware.IncUnregisteredInformers(clusterLabels, middleware.CD_STAGE_ARGO_WORLFLOW)
-	}
-	stopChan, err := impl.getCdArgoWfStopChannel(clusterLabels)
-	if err != nil {
-		return err
-	}
-	runnable := func() {
-		workflowInformer.Run(stopChan)
-		impl.logger.Infow("informer started for cd argo workflow", "clusterId", clusterInfo.Id, "clusterName", clusterInfo.ClusterName)
-	}
-	impl.asyncRunnable.Execute(runnable)
-	return nil
-}
