+++ kubewatch/pkg/informer/cluster/wire_cluster.go
@@ -2,14 +2,16 @@ package cluster
 
 import (
 	"github.com/devtron-labs/kubewatch/pkg/informer/cluster/argoCD"
-	"github.com/devtron-labs/kubewatch/pkg/informer/cluster/argoWf"
+	cdWf "github.com/devtron-labs/kubewatch/pkg/informer/cluster/argoWf/cd"
+	ciWf "github.com/devtron-labs/kubewatch/pkg/informer/cluster/argoWf/ci"
 	"github.com/devtron-labs/kubewatch/pkg/informer/cluster/systemExec"
 	"github.com/google/wire"
 )
 
 var WireSet = wire.NewSet(
 	argoCD.NewInformerImpl,
-	argoWf.NewInformerImpl,
+	cdWf.NewInformerImpl,
+	ciWf.NewInformerImpl,
 	systemExec.NewInformerImpl,
 
 	NewInformerImpl,
